fix: Move `error.rs` to protobuf
