+++ src/execution_plans/network_coalesce.rs
@@ -3,10 +3,10 @@ use crate::channel_resolver_ext::get_distributed_channel_resolver;
 use crate::common::scale_partitioning_props;
 use crate::config_extension_ext::ContextGrpcMetadata;
 use crate::distributed_physical_optimizer_rule::{NetworkBoundary, limit_tasks_err};
-use crate::errors::{map_flight_to_datafusion_error, map_status_to_datafusion_error};
 use crate::execution_plans::{DistributedTaskContext, StageExec};
 use crate::flight_service::DoGet;
 use crate::metrics::proto::MetricsSetProto;
+use crate::protobuf::errors::{map_flight_to_datafusion_error, map_status_to_datafusion_error};
 use crate::protobuf::{DistributedCodec, StageKey, proto_from_stage};
 use arrow_flight::Ticket;
 use arrow_flight::decode::FlightRecordBatchStream;
