+++ src/protobuf/errors/mod.rs
@@ -1,10 +1,11 @@
 #![allow(clippy::upper_case_acronyms, clippy::vec_box)]
 
-use crate::errors::datafusion_error::DataFusionErrorProto;
 use arrow_flight::error::FlightError;
 use datafusion::common::internal_datafusion_err;
 use datafusion::error::DataFusionError;
 use prost::Message;
+
+use crate::protobuf::errors::datafusion_error::DataFusionErrorProto;
 
 mod arrow_error;
 mod datafusion_error;
