+++ src/protobuf/errors/objectstore_error.rs
