+++ src/protobuf/errors/parquet_error.rs
